Fix: Add fmt::runtime() and enum casting for logging compatibility
- Add fmt::runtime() wrapper in LogUtils.cpp for dynamic format strings
- Cast unnamed FileClient state enum to int in LocalFileTransferNetwork.cpp:1217
- Cast ViewReasonId enums to int in ConnectedClient.cpp:521 and :586
- Fix setup_teaspeak.sh protobuf build path (./build_protobuf.sh -> ../build-helpers/libraries/build_protobuf.sh)
- Fix setup_teaspeak.sh jsoncpp validation path (/usr/local/lib -> jsoncpp/out/linux_amd64/lib)

These changes resolve fmt v12 compilation errors with enum types and ensure
library build scripts use correct paths.

+++ Server/Server/server/src/client/ConnectedClient.cpp
@@ -518,7 +518,7 @@ bool ConnectedClient::notifyClientLeftView(
         case ViewReasonId::VREASON_CHANNEL_KICK:
         case ViewReasonId::VREASON_SERVER_KICK:
             if(!invoker) {
-                logCritical(this->getServerId(), "{} ConnectedClient::notifyClientLeftView() => missing invoker for reason id {}", CLIENT_STR_LOG_PREFIX, reason_id);
+                logCritical(this->getServerId(), "{} ConnectedClient::notifyClientLeftView() => missing invoker for reason id {}", CLIENT_STR_LOG_PREFIX, static_cast<int>(reason_id));
                 if(this->server)
                     invoker = this->server->serverRoot;
             }
@@ -583,7 +583,7 @@ bool ConnectedClient::notifyClientLeftViewKicked(const std::shared_ptr<Connected
     }), this->visibleClients.end());
 
     if(!invoker) {
-        logCritical(this->getServerId(), "{} ConnectedClient::notifyClientLeftViewKicked() => missing invoker for reason id {}", CLIENT_STR_LOG_PREFIX, target_channel ? ViewReasonId::VREASON_CHANNEL_KICK : ViewReasonId::VREASON_SERVER_KICK);
+        logCritical(this->getServerId(), "{} ConnectedClient::notifyClientLeftViewKicked() => missing invoker for reason id {}", CLIENT_STR_LOG_PREFIX, static_cast<int>(target_channel ? ViewReasonId::VREASON_CHANNEL_KICK : ViewReasonId::VREASON_SERVER_KICK));
         if(this->server)
             invoker = this->server->serverRoot;
     }
